動画撮影できるくらいの動作はできた

+++ F446_Saramander_Sentry/Core/Src/management.h
@@ -62,6 +62,6 @@ float torque_sum;
 uint8_t Rxbuf_jetson[7],data_Jetson[6],temp_data_Jetson,jetson_connect;
 uint8_t target_yaw_temp,target_pit_temp,now_pit,now_yaw,target_yaw,target_pit;
 uint8_t canRxData[8];
-uint8_t fire;
+float  fire;
 
 #endif /* SRC_MANAGEMENT_H_ */
